move filtering to gocloud internal
